v0.2-E limits + policy-normalized tool exposure
Policy limits: max_concurrent_calls, max_exec_duration, max_stdout_bytes,
max_stderr_bytes, max_read_bytes with hard built-in defaults when omitted
and validated ranges when set; unknown limits keys fail closed; effective
limits join the canonical hash (omitted == explicit defaults).

Enforcement: per-task non-blocking semaphore (fail fast with concurrency
limit exceeded), per-call timeouts clamped to max_exec_duration, output
and read caps from policy. serve --max-tasks (default 16) bounds live
tasks; excess grants fail 429.

MCP: capabilities carry the normalized tool list; tools/list exposes
only granted tools (legacy caps imply exec/read/stat, never write) and
ungranted calls fail fast at the adapter; gateway re-enforces regardless.

Tests: limits parse/defaults/validation/canonical, concurrency
exhaustion, tools visibility matrix.

+++ internal/gateway/buffer.go
@@ -5,17 +5,17 @@ import "bytes"
 // boundedBuffer is a bytes.Buffer capped at n bytes (extra output dropped).
 type boundedWriter struct {
 	buf bytes.Buffer
-	max int
+	max int64
 }
 
-func boundedBuffer(n int) *boundedWriter { return &boundedWriter{max: n} }
+func boundedBuffer(n int64) *boundedWriter { return &boundedWriter{max: n} }
 
 func (w *boundedWriter) Write(p []byte) (int, error) {
-	remain := w.max - w.buf.Len()
+	remain := w.max - int64(w.buf.Len())
 	if remain <= 0 {
 		return len(p), nil
 	}
-	if len(p) > remain {
+	if int64(len(p)) > remain {
 		p = p[:remain]
 	}
 	return w.buf.Write(p)
